ambot-be/requirements.txt
fastapi
uvicorn
llama-cpp-python
sentence-transformers
faiss-cpu
numpy
pydantic
einops
